feat: increased remote performance & finalized overview & details schedule screens & improve editor screens & init info feature

+++ shared/core/domain/src/commonMain/kotlin/entities/analytics/DailyAnalysis.kt
@@ -23,7 +23,7 @@ import kotlinx.datetime.Instant
  */
 data class DailyAnalysis(
     val date: Instant,
-    val generalAssessment: Int,
+    val generalAssessment: Float,
     val numberOfClasses: Int,
     val numberOfTests: Int,
     val numberOfMovements: Int,
